+++ AtlasLoot/CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Added padding to the faction tooltip
 ### Fixed
+- Fix error on boss list frames that scroll (like TBC Factions)
 - Removed Glyph of the Headhunter from Legion Inscription Crafting
 
 ## 9.0.0 - May 26, 2025